Created new Singleton Metaclass
Previously a Singleton was a class that was inheritted by ConfigManager
and Database Classes
The Problem with the previous implementation was that even though the
same instance of the object was maintained, it still called the __init__
method repeatedly and overrid any saved state in the ConfigManager.
New implementation with a metaclass solves this.